华强：D20231012国美定制api转化接口-3710076 * 模拟容灾

+++ marketing-cs/src/main/java/com/br/marketing/service/custom/guomei/IGuoMeiDataService.java
@@ -18,7 +18,7 @@ public interface IGuoMeiDataService {
      * @return ResponseGuMeDTO
      * @author Guo Zeqiang
      */
-    Long saveTransferDataHandler(GuoMeiTransferData guoMeiTransferData);
+    int saveTransferDataHandler(GuoMeiTransferData guoMeiTransferData);
 
 
 }
